refactor(dto): 객체 노출 안되도록 dto 생성
