+++ backend/src/main/java/wooteco/retrospective/presentation/dto/attendance/AttendanceByTimeResponse.java
@@ -1,33 +1,33 @@
 package wooteco.retrospective.presentation.dto.attendance;
 
-import java.util.List;
-
 import wooteco.retrospective.application.dto.ConferenceTimeDto;
-import wooteco.retrospective.domain.member.Member;
-import wooteco.retrospective.presentation.dto.MembersDto;
+import wooteco.retrospective.application.dto.MembersDto;
+import wooteco.retrospective.presentation.dto.member.MembersResponse;
 
 public class AttendanceByTimeResponse {
 
     private Long conferenceTimeId;
-    private List<Member> members;
+    private MembersResponse members;
 
     public AttendanceByTimeResponse() {
     }
 
-    public AttendanceByTimeResponse(Long conferenceTimeId, List<Member> members) {
+    public AttendanceByTimeResponse(Long conferenceTimeId, MembersResponse members) {
         this.conferenceTimeId = conferenceTimeId;
         this.members = members;
     }
 
-    public static AttendanceByTimeResponse of(ConferenceTimeDto conferenceTime, MembersDto members) {
-        return new AttendanceByTimeResponse(conferenceTime.getId(), members.getMembers());
+    public static AttendanceByTimeResponse of(ConferenceTimeDto conferenceTime, MembersDto membersDto) {
+        MembersResponse membersResponse = new MembersResponse(membersDto.getMembers());
+
+        return new AttendanceByTimeResponse(conferenceTime.getId(), membersResponse);
     }
 
     public Long getConferenceTimeId() {
         return conferenceTimeId;
     }
 
-    public List<Member> getMembers() {
+    public MembersResponse getMembers() {
         return members;
     }
 }
